pyproject.toml
[project]
name = "ai-ides-tracker"
version = "0.1.0"
description = "Catalog and CLI to track AI IDEs & coding agents"
readme = "README.md"
authors = [{ name = "", email = "" }]
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6,<7",
  "typer[all]>=0.12,<1",
  "rich>=13,<14",
  "Jinja2>=3,<4",
]

[project.scripts]
aides = "src.cli:app"

[tool.ruff]
line-length = 100
